backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
tensorflow==2.18.0
pandas==2.2.0
scikit-learn==1.4.0
pydantic==2.6.1
